requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
anyio==3.6.2
apache-airflow==2.5.3
apache-airflow-providers-amazon==6.1.0
apache-airflow-providers-common-sql==1.3.4
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.2.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
argcomplete==3.0.5
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
beautifulsoup4==4.12.3
blinker==1.5
boto3==1.28.7
botocore==1.31.7
cachelib==0.9.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.2.35
croniter==1.3.8
cryptography==39.0.2
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.3.0
docutils==0.19
email-validator==1.3.1
exceptiongroup==1.1.1
Flask==2.2.3
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpath-ng==1.7.0
jsonschema==4.17.3
lazy-object-proxy==1.9.0
linkify-it-py==2.0.0
lockfile==0.12.2
lxml==5.3.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.5
mdurl==0.1.2
multidict==6.0.4
mypy-boto3-appflow==1.35.39
mypy-boto3-rds==1.35.72
mypy-boto3-redshift-data==1.35.51
numpy==2.1.3
packaging==21.3
pandas==2.2.3
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.0.0
ply==3.11
prison==0.2.1
psutil==5.9.4
py4j==0.10.9.7
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
pyspark==3.5.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
redshift-connector==2.1.4
requests==2.28.2
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.3.3
s3transfer==0.6.2
scramp==1.4.5
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.6
SQLAlchemy==1.4.47
SQLAlchemy-JSONField==1.0.1.post0
sqlalchemy-redshift==0.8.14
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.2.0
text-unidecode==1.3
typing_extensions==4.5.0
tzdata==2024.2
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.15
watchtower==2.0.1
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.8.2
